Fix branch

+++ src/main/java/ar/edu/unlam/tallerweb1/delivery/DatosEntrada.java
@@ -6,7 +6,7 @@ import ar.edu.unlam.tallerweb1.domain.usuario.Usuario;
 public class DatosEntrada {
 	private Funcion funcion;
     private Usuario usuario;
-    private Long cantidad;
+    private Integer cantidad;
 
     
 	public Funcion getFuncion() {
@@ -21,10 +21,10 @@ public class DatosEntrada {
 	public void setUsuario(Usuario usuario) {
 		this.usuario = usuario;
 	}
-	public Long getCantidad() {
+	public Integer getCantidad() {
 		return cantidad;
 	}
-	public void setCantidad(Long cantidad) {
+	public void setCantidad(Integer cantidad) {
 		this.cantidad = cantidad;
 	}
